Hehe oops crash when no reply :3

+++ index.js
@@ -5,7 +5,7 @@ const client = new Client({
 	intents: [
 		GatewayIntentBits.Guilds,
 		GatewayIntentBits.GuildMessages,
-		GatewayIntentBits.MessageContent,
+		GatewayIntentBits.MessageContent
 	]
 });
 
@@ -18,10 +18,16 @@ client.on("messageCreate", async (msg) => {
 		if (msg.author.id in users) {
 			if (users[msg.author.id].includes(msg.channelId)) {
 				const reply = msg.reference;
-				
+
+				if (reply === null) {
+					return;
+				}
+
 				if (reply.channelId === msg.channelId) {
 					if ("messageId" in reply) {
-						const messages = await msg.channel.messages.fetch({ after: reply.messageId });
+						const messages = await msg.channel.messages.fetch({
+							after: reply.messageId
+						});
 
 						await msg.channel.bulkDelete(messages, true);
 					}
